backend/pyproject.toml
[project]
name = "rhythmlink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-jazzmin>=3.0.1",
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "environ>=1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "xlrd>=2.0.2",
]
